LayoutTests/http/wpt/dom/inline-module-script-timing-expected.txt

CONSOLE MESSAGE: in classic script => document.readyState = loading
CONSOLE MESSAGE: [module script] in module script => document.readyState = interactive
CONSOLE MESSAGE: [module script] in promise_test => document.readyState = interactive
CONSOLE MESSAGE: [module script] in Promise() created
CONSOLE MESSAGE: [module script] in window.onload
This timing is critical to run with wpt testharness.js

PASS DOM loading state verification (module script)
